+++ src/cases/services/index.ts
@@ -3,4 +3,4 @@ import { SearchFilterService } from './search-filter.service';
 export const services: any[] = [SearchFilterService];
 
 export * from './search-filter.service';
-export * from './decentralised-redirect.service';
+export * from '../../decentralisation/decentralised-redirect.service';
